refactor(tests): Tier 3 batch 1 — split mcp/helpers_test.go and buildgraph/generator_test.go
Split cmd/mcp/helpers_test.go (418 lines) into three focused files by concern:
config overrides, output/result helpers, and build entry/cost logic.

Split internal/buildgraph/generator_test.go (345 lines) by extracting all
assertion and lookup helpers into generator_helpers_test.go, leaving
generator_test.go as a clean list of TestGenerate_* functions only.

No production code changes. All 53 test packages pass.
